third_party/WebKit/LayoutTests/fast/dom/Document/createProcessingInstruction-invalid-target.html

<!doctype html>
<html>
<head>
<meta http-equiv="Content-Type" content="text/html; charset=utf-8">
<script src="../../../resources/js-test.js"></script>
</head>
<body>
<script>
[
    // Invalid first Char
    '0ascii',
    '.Ascii',
    '-Ascii',
    ' ascii',
    '	ascii',
    '֑sc.ii',
    '⃣scii',
    'िascii',
    'ʳascii',
    // Invalid other char
    'asc i',
    'asc	i',
    'asciiⅦ'
].forEach(function(target)
{
    shouldThrow("document.createProcessingInstruction('" + target + "', '')");
});
</script>
</body>
</html>
