third_party/blink/web_tests/fast/dom/Document/createProcessingInstruction-invalid-target-expected.txt

PASS document.createProcessingInstruction('0ascii', '') threw exception InvalidCharacterError: Failed to execute 'createProcessingInstruction' on 'Document': The target provided ('0ascii') is not a valid name..
PASS document.createProcessingInstruction('.Ascii', '') threw exception InvalidCharacterError: Failed to execute 'createProcessingInstruction' on 'Document': The target provided ('.Ascii') is not a valid name..
PASS document.createProcessingInstruction('-Ascii', '') threw exception InvalidCharacterError: Failed to execute 'createProcessingInstruction' on 'Document': The target provided ('-Ascii') is not a valid name..
PASS document.createProcessingInstruction(' ascii', '') threw exception InvalidCharacterError: Failed to execute 'createProcessingInstruction' on 'Document': The target provided (' ascii') is not a valid name..
PASS document.createProcessingInstruction('	ascii', '') threw exception InvalidCharacterError: Failed to execute 'createProcessingInstruction' on 'Document': The target provided ('	ascii') is not a valid name..
PASS document.createProcessingInstruction('֑sc.ii', '') threw exception InvalidCharacterError: Failed to execute 'createProcessingInstruction' on 'Document': The target provided ('֑sc.ii') is not a valid name..
PASS document.createProcessingInstruction('⃣scii', '') threw exception InvalidCharacterError: Failed to execute 'createProcessingInstruction' on 'Document': The target provided ('⃣scii') is not a valid name..
PASS document.createProcessingInstruction('िascii', '') threw exception InvalidCharacterError: Failed to execute 'createProcessingInstruction' on 'Document': The target provided ('िascii') is not a valid name..
PASS document.createProcessingInstruction('ʳascii', '') threw exception InvalidCharacterError: Failed to execute 'createProcessingInstruction' on 'Document': The target provided ('ʳascii') is not a valid name..
PASS document.createProcessingInstruction('asc i', '') threw exception InvalidCharacterError: Failed to execute 'createProcessingInstruction' on 'Document': The target provided ('asc i') is not a valid name..
PASS document.createProcessingInstruction('asc	i', '') threw exception InvalidCharacterError: Failed to execute 'createProcessingInstruction' on 'Document': The target provided ('asc	i') is not a valid name..
PASS document.createProcessingInstruction('asciiⅦ', '') threw exception InvalidCharacterError: Failed to execute 'createProcessingInstruction' on 'Document': The target provided ('asciiⅦ') is not a valid name..
PASS successfullyParsed is true

TEST COMPLETE
